download count is sent as a float because CF

+++ src/api.rs
@@ -114,7 +114,7 @@ pub struct File {
 	#[serde(rename = "fileLength")]
 	pub file_length: i64,
 	#[serde(rename = "downloadCount")]
-	pub download_count: f64, // CF WHY
+	pub download_count: i64,
 	#[serde(rename = "downloadUrl")]
 	pub download_url: String,
 	#[serde(rename = "gameVersions")]
@@ -148,7 +148,7 @@ pub struct Mod {
 	pub summary: String,
 	pub status: ModStatus,
 	#[serde(rename = "downloadCount")]
-	pub download_count: i64,
+	pub download_count: f64, // cf...
 	#[serde(rename = "isFeatured")]
 	pub is_featured: bool,
 	#[serde(rename = "primaryCategoryId")]
